requirements.txt
flask
flask-cors
llama-cpp-python #--prefer-binary for installation 
torch
huggingface_hub
transformers